Consolidate side policy panels #765
